Add support for embedded expressions in template strings, for better autocompletion in current TypeScript editors.
Within ES6 template strings, '${' unambiguously refers to the beginning of an embedded expression. Because the actual value of the expression is evaluated at run-time, this currently has no meaning (and will fail) when inside a React.jsx() call, which is essentially a compile-time directive. Thus, we can safely denote "${}" to mean the same thing as "{}", and replace all occurrences in this way immediately before invoking the React transform.